requirements.txt
pdfplumber==0.11.4
pandas==2.2.2
numpy==1.26.4
regex==2024.9.11
python-dotenv==1.0.1
sentence-transformers==3.0.1
transformers==4.42.4
torch>=2.2.0
scikit-learn==1.5.1
streamlit==1.37.0
tqdm==4.66.4
pinecone-client==5.0.1


